\name{MFSV}
\alias{MFSV}
\title{Doppler vs cross-sectional echocardiography}
\description{
  The desults of a study by Zhang et al. (1986), measuring transmitral volumetric flow (MF) by Doppler echocardiography and left ventricular stroke volume (SV) by cross-section echocardiography in 21 patients without aortic valve disease.
}
\arguments{
  \item{MFSV}{A dataset with 21 entries and 2 variables:}
  \item{$MF}{transmitral volumetric flow measured by Doppler echocardiography.}
  \item{$SV}{left ventricular stroke volume measured by cross-sectional echocardiography.}
}
\details{
  A dataset in which both methods are measuring the same flow rate and the true flow rate is unknown.  This dataset is designed to be analyzed using the Bland Altman plot.
}
\seealso{ \code{\link{BlandAltman}}, \code{\link{agpop}} }
\keyword{datasets}
