\name{myHist}
\alias{myHist}
\title{Drawing simple histograms within a plotting area}
\description{
  This function is an internal function used to draw histograms within a plotting area.
}
\usage{
  myHist(xl, yb, xr, yt, hDat, breaks = NULL, sample = TRUE,CIs = FALSE, xlab = FALSE, label.bars = FALSE, right = TRUE)
}
\arguments{
  \item{xl,yb,xr,yt}{}
  \item{hDat}{}
  \item{breaks}{}
  \item{sample}{}
  \item{CIs}{}
  \item{xlab}{}
  \item{label.bars}{}
  \item{right}{}
}
\details{

}
\value{

}
\note{

}
\author{ Mohamed Abdolell <mohamed.abdolell@dal.ca> and Sam Stewart <samstewart11@gmail.com> }
\references{ This plot was designed for a course by Mohamed Abdolell }
\examples{ 

}
\keyword{distribution}
